feat(workerproto): tell the worker which assignments are parked
Leaving the worker unable to learn that an attempt is parked was not an
acceptable resting place. A worker that collected a parked attempt made that
attempt fail rather than resume, and a task parked on a wait must not be
considered complete, verified or failed because its turn ended.

The coordinator now states, in the snapshot exchange the worker already makes,
which of that worker's claimed assignments are parked on a live task-bound
wait, with the attempt revision the park is fenced on. The worker is told and
never asks, so the restricted worker protocol stays restricted: no read of
coordinator state is added, and the statement is scoped to the addressed
worker's own assignments.

The statement is complete rather than incremental, because an incremental one
cannot say "no longer parked" without another message to lose, and it is
durable, because a worker that restarts mid-wait runs its own reconcile loop
before its next exchange and would otherwise collect. A report flagged as not
reporting at all is an older coordinator and leaves the previous behaviour
alone. A report that is stale, or that describes another execution of the same
assignment, is unknown rather than negative, and unknown defers collection.
An entry naming an attempt revision behind one already applied does not
overwrite it: reports can overtake each other on a retried exchange.

The coordinator's refusal is unchanged and remains the authority. This makes
the normal path graceful instead of a failure the coordinator has to catch.

Following the contract update: ExpectedRevision is now int64 to match
Attempt.Revision, and mixing each and all on one attempt is defined where the
type is declared as any each settlement waking the attempt.

+++ internal/store/sqlite/task_wait.go
@@ -124,7 +124,7 @@ func (s *Store) RegisterTaskWait(ctx context.Context, request domain.TaskWaitReg
 	if attempt.Progress.Terminal() {
 		return wait, domain.TaskWaitTerminalRefusal(attempt.Progress)
 	}
-	if uint64(attempt.Revision) != request.ExpectedRevision {
+	if attempt.Revision != request.ExpectedRevision {
 		return wait, fmt.Errorf("%w: attempt %q is at revision %d, not %d",
 			domain.ErrTaskWaitStaleRevision, attempt.ID, attempt.Revision, request.ExpectedRevision)
 	}
